Optimize balance-related queries with a cache (#2383)
## Linked Issues/PRs
Closes https://github.com/FuelLabs/fuel-core/issues/1965

## Description
1. The client uses the off-chain database metadata to tell if it can use
the optimized queries or fall-back to the legacy ones.
2. When client starts the metadata is either:
1. Kept intact in-case genesis exists (meaning - no optimized queries
are available)
* this stems from the fact that the pre-existing DB contains some coins
already and we'll **not** be re-indexing them in order to create the
lookup indexes
* in this mode of operation the "old" way of getting balances is used as
confirmed in logs:
          * `worker_service: 607: Balances cache available: false`
* `query::balance: 100: Querying balances without balances cache
owner=53a9c6a...`
    2. Initialized with `V2` if genesis is missing
* this means that while re-syncing the DB the balance indexes are going
to be created and could be used to satisfy the requests
          * `worker_service: 607: Balances cache available: true`
* `query::balance: 151: Querying balances using balances cache
owner=53a9c6a...`

Technical considerations:
* Metadata is extended with `V2` which now contains
`indexation_availability`. This is a set that holds the available
indexations (currently only `Balances`)
* New databases are added: `CoinBalances` and `MessageBalances` that
keep the coin and message balances respectively
   * `CoinBalances` stores the balances per `owner` and `asset_id`
* `MessageBalances` stores balances per `owner`, separately for
retryable and non-retryable messages
* In case balances indexation is available (see "Description" above):
* these databases are updated in the GrapQL API worker service when
processing one of these events: `MessageImported`, `MessageConsumed`,
`CoinCreated`, `CoinConsumed`
* `balance()` and `balances()` queries use these databases to satisfy
the query

## Checklist
- [X] New behavior is reflected in tests

### Before requesting review
- [X] I have reviewed the code myself

---------

+++ tests/tests/chain.rs
@@ -170,11 +170,11 @@ async fn network_operates_with_non_zero_base_asset_id() {
         .expect("transaction should insert");
 
     // Then
-    let expected_fee = 1;
+    let expected_fee = 1_u128;
     assert!(matches!(result, TransactionStatus::Success { .. }));
     let balance = client
         .balance(&owner, Some(&new_base_asset_id))
         .await
         .expect("Should fetch the balance");
-    assert_eq!(balance, amount - expected_fee);
+    assert_eq!(balance, amount as u128 - expected_fee);
 }
